Revert "Refactor risk dashboard modules into layered packages"
