--- risk_management/core/__init__.py
@@ -1,25 +1,0 @@
-"""Core domain and policy utilities for risk management."""
-
-from . import domain
-from .policies import (
-    PolicyActionState,
-    PolicyEvaluation,
-    PolicyEvaluationResult,
-    PolicyEvaluator,
-)
-from .snapshots import (
-    evaluate_alerts,
-    normalise_policy_summary,
-    parse_snapshot,
-)
-
-__all__ = [
-    "domain",
-    "evaluate_alerts",
-    "normalise_policy_summary",
-    "parse_snapshot",
-    "PolicyActionState",
-    "PolicyEvaluation",
-    "PolicyEvaluationResult",
-    "PolicyEvaluator",
-]
