--- risk_management/core/domain/__init__.py
@@ -1,25 +1,0 @@
-"""Domain package for risk management models and value objects."""
-
-from .models import (
-    Account,
-    AlertThresholds,
-    Order,
-    Position,
-    Scenario,
-    ScenarioAccountImpact,
-    ScenarioResult,
-    ScenarioShock,
-    ScenarioSymbolExposure,
-)
-
-__all__ = [
-    "Account",
-    "AlertThresholds",
-    "Order",
-    "Position",
-    "Scenario",
-    "ScenarioAccountImpact",
-    "ScenarioResult",
-    "ScenarioShock",
-    "ScenarioSymbolExposure",
-]
