+++ risk_management/domain/__init__.py
@@ -1,5 +1,5 @@
-"""Backward compatible accessors for risk management domain models."""
+"""Domain package for risk management models and value objects."""
 
-from ..core.domain import Account, AlertThresholds, Order, Position
+from .models import Account, AlertThresholds, Order, Position
 
 __all__ = ["Account", "AlertThresholds", "Order", "Position"]
