--- risk_management/io/__init__.py
@@ -1,5 +1,0 @@
-"""Client and fetcher implementations used by the risk management services."""
-
-from .account_clients import AccountClientProtocol, CCXTAccountClient
-
-__all__ = ["AccountClientProtocol", "CCXTAccountClient"]
