+++ risk_management/liquidity.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+"""Utilities for analysing order-book liquidity and slippage."""
+
 from typing import Any, Iterable, List, Mapping, MutableMapping, Optional, Sequence, Tuple
 
 __all__ = ["normalise_order_book", "calculate_position_liquidity"]
