+++ risk_management/services/__init__.py
@@ -1,6 +1,12 @@
-"""Service package exports for risk management."""
+"""Service layer helpers for the risk management dashboard."""
 
 from .performance_repository import PerformanceRepository
+
+__all__ = ["PerformanceRepository"]
+
+"""Service abstractions for risk management workflows."""
+
 from .risk_service import RiskService, RiskServiceProtocol
 
-__all__ = ["PerformanceRepository", "RiskService", "RiskServiceProtocol"]
+__all__ = ["RiskService", "RiskServiceProtocol"]
+
